refactor: Extract LLM models and constraints into a dedicated module
